refactor(lexer): move convert_to_positional_params to pgls_lexer
The named-parameter normalization is purely lexical and is about to be needed by the statement splitter as well. Moving it to pgls_lexer keeps the splitter and the workspace in agreement about what counts as parsable SQL.

Pure move: no signature, body or comment changed.
